+++ src/main/java/net/frozenblock/wilderwild/registry/WWBiomes.java
@@ -53,7 +53,7 @@ import net.minecraft.resources.ResourceKey;
 import net.minecraft.world.level.biome.Biome;
 import org.jetbrains.annotations.NotNull;
 
-public final class WWWorldgen {
+public final class WWBiomes {
 	// Main Biomes
 	public static final ResourceKey<Biome> CYPRESS_WETLANDS = CypressWetlands.INSTANCE.getKey();
 	public static final ResourceKey<Biome> OASIS = Oasis.INSTANCE.getKey();
@@ -92,7 +92,7 @@ public final class WWWorldgen {
 	public static final ResourceKey<Biome> OLD_GROWTH_DARK_FOREST = OldGrowthDarkForest.INSTANCE.getKey();
 	public static final ResourceKey<Biome> SNOWY_OLD_GROWTH_PINE_TAIGA = SnowyOldGrowthPineTaiga.INSTANCE.getKey();
 
-	private WWWorldgen() {
+	private WWBiomes() {
 		throw new UnsupportedOperationException("WWWorldgen contains only static declarations.");
 	}